pyproject.toml
[project]
name = "ptb-starlette-starter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "jinja2==3.1.5",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "python-telegram-bot[rate-limiter]==22.4",
    "sqlalchemy>=2.0.43",
    "starlette==0.45.1",
    "uvicorn==0.34.0",
]
